Added functionality to go through a CSV file and calculate the sum of all spendings for that day
Added a function that reads the CSV file based on the file path and returns the contents of the file.

Added function that finds a column based on the column name and returns its values.

Added function that calculates the sum of all the values for a specific column, prints and returns the result.